Remove :-free shorthand (part 1)
Summary:
TARGETS files dependencies contain a colon-free shorthand, e.g.
  deps = [ '@/common/base' ],  # expands to '@/common/base:base'
This shorthand is confusing to engineers, both new and old alike, and so
should be removed.
This shorthand does not exist in buck, and so needs to be removed.

All TARGETS files fixed up via the command
  find . -name TARGETS | grep -vE "^\./(third-party-buck|third-party2)/" | xargs  sed -i "s,\(['\"][@/]/\)\([^'\":]*/\)*\([^'\":/.][^'\":/.]*\)\(['\"]\),\1\2\3:\3\4,g"

The full original diff seems to be too big for phabricator. I've
run ##git checkout HEAD~1 [n-z]*## to halve the size of this diff.

Once this diff is approved, I'll locally revert all the changes, re-run
the codemod command, and then land immediately.

Once this diff lands, I'll wait a few days, then do a second pass that
fixes new errors and also prevents this type of error from being
re-added. This will cause much fewer merge conflicts for other engineers
with in-flight diffs.

Differential Revision: D2329305

fbshipit-source-id: 181c6dbb97f14002786086b941d6696741d96f6d