Do not leak HTTP response strings, FDs
Summary:
It's not the greatest that our scheduler OOMs every few hours -- the resulting downtime is annoying, and with the concurrent TW issues, the unpleasantness compounds.

I'm not actually sure if this fixes our older "periodic FD exhaustion bug", but it's certainly true that we will now be leaking fewer FDs. If there are others, we'll find out.

Differential Revision: D2479226

fbshipit-source-id: 6a92844c7afce74e68b949b3e6b3ddcbf196ea4c