Get rid of killed task status filters
Summary:
Currently, Bistro's task status handling is divergent depending on whether or not Bistro had sent SIGTERM/SIGKILL to the task before it exited.

(1) Not signaled:
(1a) Empty status file: errorBackoff("Failed to read a status")
(1b) Non-empty status: Parse the status, use it as-is.

(2) Signaled:
(2a) Empty status: incompleteBackoff("Task killed, no status returned")
(2b) Non-empty status: Parse the status, but coerce it to a different value as per KilledTaskStatusFilter.

This is racy, in the sense that the task could have exited normally just before we signaled it. This means that, depending on whether Bistro tries to waitpid() or kill() first, the status semantics will differ -- even though the task exited of its own accord in both cases.

This non-determinism feels broken, and ideally (1) and (2) would behave identically. On the other hand, there really can be several different intentions for killing a task:

(i) Act as if the task just exited on its own -- gives the task an opportunity to handle a SIGTERM and print a status. Less sophisticated tasks would be defaulted to errorBackoff("Failed to read a status") as per 1a.

(ii) Preempt the task, so that others can run. Preemption should be maximally transparent, so unless the task handles the signal and expliclty prints a status, treat the status as incompleteBackoff("Task killed, no status returned") to protect the retry counter.

This difference in intention explains an essential distinction between (1a) and (2a). This diff keeps only this distinction, and removes the remaining complexity.

The proposed simplified logic is:

- If the task outputs a non-empty status, always use that (no coercion).
- If the task outpus nothing, use "errorBackoff" for "voluntary exit" situations; use "incompleteBackoff" for "killed". This is racy -- failing to appropriately decrement a retry counter -- when the task voluntarily exits, and Bistro simultaneously runs kill() before wait().

In exchange for simplicity, we give up the status coercion feature, and the ability to let preempted tasks restart immediately with no backoff period.

Reviewed By: alexwwyy

Differential Revision: D2285589

fbshipit-source-id: a5139d5c172f53bcc8567917856cfd7a73bf2f13