Refactor to permit pausing/resuming reading from a pipe
Summary:
I'm stacking this on top of D2079185 since it's a fairly distinct iteration on what was a pretty good concept.

This changes the API so that we can temporarily suspend receiving READ notifications, and then resume them later.

One motivation for this change is that Bistro needs to be able to rate-limit its incoming pipes.
Another motivation is that `Subprocess` had a sometimes-used `enableNotifications()` facility of a similar intent.

Differential Revision: D2138898

fbshipit-source-id: 34ff8783b94be67980a01f8c2f738cf761a20c6d