Wait for / signal a subprocess using an EventBase
Summary:
This is a counter-part to D2079185 for IO. I'm pretty happy with this API.

I'm not a fan of the argument that this is better done with 'signalfd', since it's Linux-only and forces SIGCHLD to be blocked process-wide (which is hard for a library to impose). The "poll - delay - poll" method is plenty good enough for most use-cases, and is portable and robust.

I think this is, overall, folly-appropriate, but somebody else should want this before I commit to going through all this trouble.

The "race waitpid vs signal" unit test is probably the most amusing part of this diff (but the API merits more of your attention).

Reviewed By: vii, barillari

Differential Revision: D2082184

fbshipit-source-id: 3749efb0466452bdd799ee3c558569e4954c6874