Exit initial wait early if workers report a consensus
Summary:
This is the money shot:
 - Actually track the current worker set ID on the scheduler, and pass it to the workers.
 - If all the workers' `initialWorkerSetID_`s are the same, and match the scheduler's `nonMustDieWorkerSetID`, we exit initial wait early.

However, we are not yet done, since as-written, new workers are able to start running tasks immediately, whether or not they are included in a consensus of prior workers, and remote/README.worker_set_consensus (prior diff) explains why that's a bad idea.

So the remaining diffs will make `consensus_permits_becoming_healthy` actually do something.

Push-blocking tests broken:

Differential Revision: D2657580

fbshipit-source-id: b020297175cb202ac0a70ff454bbb80471d59f18