Prune RemoteWorkers history to keep a low runtime complexity
Summary:
This is an easily separated add-on to the label propagation diff for worker set ID consensus.

Without the history pruning implemented here, the typical update step would take time proportional to the number of history updates, which means that we'd gradually slow down over time. With history pruning, the length of an update step is (normally) proportional to the number of workers. This is much better-behaved, and easy to implement, so it seem worth doing.

Push-blocking tests broken:

Differential Revision: D2657591

fbshipit-source-id: bc10276c95d3404d24bad5243ad8c62e956cf2aa