Iterate over cgroups to wait for, and/or kill, tasks
Summary:
Similarly like D2702652, this looks at some files & directories. If all the cgroups become empty, we decide the task is done. This is a much stronger form of "wait for the whole task" than the current fd-based canary.

After massaging the `freezer` cgroup into the right state, this can also SIGKILL all processes in it.

The details are in the docs, but in essence, this avoids this pernicious race between:

 - I read some cgroup PIDs
 - Some of these PIDs exit, and get recycled
 - I signal the PIDs, but I just killed the wrong processes.

The trick is to tell the `freezer` subsystem to stop giving **any** CPU cycles to any processes in our group. Then they can neither forkbomb nor exit. After the freeze, we can SIGKILL safely. Once they thaw, they die.

Like D2702652, this is fairly independent of Bistro and can be reused.

contbuild is always broken

Differential Revision: D2702656

fbshipit-source-id: 21e2278fae22b572a82385980e0eaf0b8fe6d032