Add a way to manually abort initial wait
Summary:
Sometimes one is doing scheduler maintenance, where the workers are also restarted, and thus one is certain that initial wait is unnecessary.

This flag is helpful in this case to minimize downtime.

Note that misusing this flag can cause your scheduler to double-start tasks, and later crash with with something like "CHECK: task running twice".

Reviewed By: barillari

Differential Revision: D2753105

fbshipit-source-id: 427e199e95ef1f9724687772854b25457202a052