Rename scheduler binary to `bistro_scheduler`
Summary: This is a much clearer naming setup, I've wanted to do this for a while. Since the new builds of the scheduler & worker are fairly backwards-incompatible, this is also a good litmus test for whether the person pushing a new Bistro package has seen my announcement in the users group (coming soon).

Reviewed By: pmenglund

Differential Revision: D2754490

fbshipit-source-id: 454817d190a9528880d9713cb0bf24c4561592b5