Make cpu.shares always be >= 2
Summary:
Linux gives you an error if you try to set `cpu_shares` values under 2. So, Bistro shouldn't do that. This affected the FAIR tiers in December, so I made this fix, but never actually sent it out until now.

In particular, this means that we can re-enable CPU fairness enforcement for FAIR tiers (though people don't seem to care?).

Reviewed By: soumith

Differential Revision: D3136897

fbshipit-source-id: f0813d968dd957ac3083b1dbb583f1b797e1da64