Fix replica handling in the new optimized scheduler
Summary:
vladsimonenko, committing this depends on D3136954, so consider either reviewing it, or bugging agoder to do so :)

While D2702649 made the scheduler a lot faster, it broke replica handling by allocating a separate resource slot, and separate `JobNode` for each replica, allowing multiple identical tasks to start when replicas are in use (which predictably crashes the scheduler).

This sacrifices a tiny bit of performance to ensure that all replicas share the same resource slots, and only one `JobNode` is created.

Differential Revision: D3162163

fbshipit-source-id: 064abb44f78f3c9f12724345eb8c979c0d05f749