Fix crash in getJobLogs
Summary:
Before this change, Bistro would crash if multiple getJobLogs requests hit the same HTTPMonitor thread in rapid succession.

The reason was that getJobLogs called `loopForever` on the current thread's EventBase ... but `event_base_loop()` is not reentrant.

If the requets were not closely spaced, then both nested `loop`s would exit, and the whole thing would seemingly recover.

But with the right timing, the second request ran from inside the outer loop while its state was already corrupted. This led to heap corruption and weird crashes.

The only substantive change in this diff is replacing `getEventBase` by a new `folly::EventBase` on the stack.

Also see D3408155 which makes these heisenbugs easier to prevent.

Reviewed By: drej82

Differential Revision: D3408512

fbshipit-source-id: 9a0d4cd1a4487a09fb42eb2a2ecf344e423c1b2c