Introduce a `workerLost` status that separates the configured backoff duration from the effective one
Summary:
This is used for fixing a scheduler bug. The root cause was pretty simple -- first, a worker would get lost. We would then start new instances of its tasks before we were reasonably confident that the old had been terminated.

On occasion, the old tasks would manage to return exit statuses, triggering a `CHECK` in the scheduler designed to detect double-started tasks.

The natural fix is to extend the backoff durations for tasks of lost workers. This diff allows doing this efficiently.

See a follow-up diff "Mandatory extended backoff for tasks of lost workers" to see this new status get used.

Reviewed By: agoder

Differential Revision: D3475381

fbshipit-source-id: 7437db04a651846461838e336fa6dacac3fd07b8