Don't crash while trying to update a non-running task status with an "overwriteable" one
Summary:
In a rare race, a task can complete **after** `updateState` decided the task got lost, but before `applyUpdate` managed to update TaskStatuses with the "task lost" status.

We should not crash here. The test verifies that we hit the log message that I used to replace the `CHECK`, and that we don't crash. In this sense, it confirms my root-cause analysis of the CHECK failure.

This is the **second** of the two scheduler crashes reported on #11331786. The first is fixed separately.

I also introduce a beefier test harness for testing such subtle timing issues. It makes the tests in "Mandatory extended backoff for tasks of lost workers" much easier to write, too.

Reviewed By: agoder

Differential Revision: D3475389

fbshipit-source-id: 67aacf534330fab6552f3afbd6a761302c6d3f1c