Mandatory extended backoff for tasks of lost workers
Summary:
See the detailed in-code comments. In short, the scheduler was crashing on the `CHECK` in `TaskStatusSnapshot` below (see its new, extended comment), because the scheduler had, incorrectly, started two copies of the same task.

Triggering this required a small coincidence -- first, a worker had to get lost. Since the scheduler would just use the task's regular backoff at this point, we would then be able to start new instances of its tasks before we were reasonably confident that the old had been terminated.

On occasion, the old tasks would manage to return exit statuses after the new one started, triggering this `CHECK`.

This diff fixes all this by forcing extended backoff durations for tasks of lost workers. See the in-code docs for the caveats.

Reviewed By: agoder

Differential Revision: D3475398

fbshipit-source-id: 4abaa21acbeca2f6503ea71fb109d0962aa4ac53