Fix scheduler crash when shard moves between hosts & wins a race
Summary:
Um, when I added per-host worker pools, I evidently assumed that the scheduler will always accept an incoming heartbeat. This is not true.

It can happen normally that the worker will win the race -- going down and coming back up on a new host before the scheduler considered the original worker lost. All that needs to happen is that the worker process is failed over to free up the host that initially contained it.

Reviewed By: agoder

Differential Revision: D3473552

fbshipit-source-id: 94647d230ba26c14f6cbf92ece87ca0c0c3b38fb