Default version to -1
Summary:
Bistro expects the job version ID to be unset (or -1) when a new job is being added. This allows ConfigLoaders to ensure that add & update events can never race and silently clobber each other.

Unfortunately, the deprecated Thrift API defaulted the version ID to 0, and `saveJob` was failing because the scheduler expected to be making an update, when the actual intent was to make a brand-new job.

Differential Revision: D3749950

fbshipit-source-id: 9f59aaeff1cd269b5531adcbc863967aa9ff457a