fbcode_builder: Generic DSL for continuous integration builds
Summary:
This is the conceptual descendant of `emit-dockerfile.py` that I wrote for Proxygen.

There were some serious issues I found with trying to iterate on that, including:
 - hard to support local repos
 - hard to support building as non-root, and installing into a prefix
 - shell quoting issues
 - verbose configuration for each new service
 - consistent / reproducible builds from a single git hash (coming soon)

`fbcode_builder` is a refactor/rewrite of the same idea that addresses all these issues and more.

Moreover, it is packaged up in such a way that many Facebook open-source projects can lean on it for builds. I'll try to migrate Proxygen soon.

Differential Revision: D4424566

fbshipit-source-id: 0b1413c9608d92921914e4a4dc17ac5c760f9666