Use ShipIt dependency hashes for a consistent build
Summary:
Previously, when building e.g. Proxygen, we would check out the master branch of dependencies like Folly. This could result in version shear, and the odds of it would go up as you iterated on your branch, while the dependencies moved forward. This degrades our developer experience.

The solution is to actually build from the same point in Facebook's monorepo.

As implemented, this is still imperfect, because the in-tree dependency hashes under `build/deps/github_hashes` are not currently updated atomically. However, on balance, this predictability should still be better -- and if I recall sdwilsh's explanation correctly, the atomicity of dependency commits could be fixed if it really impacts many people.

This will also improve cache coherence when you're using ccache for iteration, see D5648461.

(Also includes some extra "ideas" comments, and a whitespace fix.)

Reviewed By: sdwilsh

Differential Revision: D5648388

fbshipit-source-id: 1baa68ff3ef2e204a6cf5f5297e52514a3b62d21