Standardize the representation of boost, gflags, and sparsehash
Summary:
One step toward the goal of eliminating deprecated rules `common/base:base` and `common/base:base_deprecated` is the removal of "boost", "gflags", and "sparsehash" from their `external_deps`.
Because a high percentage of TARGETS indirectly depend on common/base:base, this requires that external_deps be specified by the TARGETS that actually use the external libraries rather than
relying on the `external_deps` of `common/base:base`.  foundation/dependency_management/ensure-explicit-dependencies.sh will be used in a subsequent diff to add the required
`external_deps` entries to those TARGETS files, and it depends on there being a single representation of "boost", "gflags", and "sparsehash" in TARGETS in order to avoid adding duplicate entries.

Reviewed By: meyering, Orvid

Differential Revision: D6015074

fbshipit-source-id: 0c3db53c4e7f73e9d738275f5fbe3f2b8d392324