buck macros: add a `no_undefined` parameter for C/C++ libraries
Summary:
This is part of of enabling `-Wl,--no-undefined` for shared library links, so
that we see undefined symbol errors at the time we link the shared library,
rather than when linking top-level binaries (or at runtime).

This diff only adds the parameter, so that a subequent codemod can whitelist
existing rules which fail to link with the above flag.

Reviewed By: yfeldblum

Differential Revision: D6427723

fbshipit-source-id: b341f5ef114c821a995fc5e1088c58dd7a5a3118