add float-cast-overflow UBSAN suppressions
Summary:
I'm going to enable UBSAN float-cast-overflow check in fbcode. It detects conversion to, from, or between floating-point types which would overflow the destination.

This diff adds suppressions to the existing violations. It is mechanically generated by a script that parses test logs and
generates the suppression and runs clang-format on the changed lines.

Now it uses TT123 as a task number, but once accepted I'll regenerate this diff with tasks.

Reviewed By: yfeldblum

Differential Revision: D6913158

fbshipit-source-id: 512d55f267ee2a036038bbc65245138a4e49c2ae