fbcode_builder: getdeps: hook into nbtd
Summary:
This is "Plan C" for NBTD which is to delegate the determinatoring
to getdeps entirely, in order to avoid duplicating dependency information.
Manually maintaining the dependency information in NBTD is fragile and prone
to breakage, and automating the dependency map inside NBTD is forbidden.

This adds `getdeps.py determinator`, a command that obtains the list of changed
files for the current commit and then emits the set of sandcastle jobs that
should be run to exercise the change.

Reviewed By: zertosh

Differential Revision: D14719886

fbshipit-source-id: 45c905c0a3f4edba18fb991bd5208e52dcdfbfff