fbcode_builder: getdeps: fboss: extend iproute2 builder to copy lib/*
Summary:
D15588809 introduced manifest and customer builder for iproute2. While the
Iproute2Builder copies include/* from build to install directory, it does not
copy lib/* which contains the libnetlink.a

As a result, the build system looks up libnetlink.a on local machine (which it
may not find).

Fix it by explicitly coping lib/* as well.

Reviewed By: wez

Differential Revision: D15708835

fbshipit-source-id: 3f5906eb48ce982b6e887cca0d3c11314425beab