getdeps: ensure that INST/lib/*.dll are found in PATH on windows
Summary:
We were troubleshooting an issue with the eden tests on windows
where the boost dlls where not being found during gtest discovery.

When we compute the environment, we were only including INST/bin in the
PATH on windows.  On Windows, the dlls are searched for in the PATH, and
since boost installs those into its `lib` dir we were missing those.

This diff causes `lib` dirs to get added to PATH on windows in the same
manner that we would add them to `LD_LIBRARY_PATH` on linux.

Reviewed By: pkaush

Differential Revision: D17694542

fbshipit-source-id: 143a907e6d30d8c12360caa43c8d9c26ff8c88c6