remove uses of Synchronized::operator->
Summary:
Partially remove references to Synchronized::operator-> and replace them with wlock() and rlock() calls, depending on context.

I opted for preserving behavior rather than switching from wlock() to rlock(), even if the existing logic was buggy or suboptimal.

This task proved to be larger than it make sense for a human to do, and the remainder of the conversions should be done with an automated codemod.

Reviewed By: yfeldblum

Differential Revision: D17799649

fbshipit-source-id: 7fc2a34f73db5bf94a97d898c66371f7d70d1909