Migrate uses of Thrift __isset to _ref() accessors
Summary:
Thrift used to generate `__isset` fields for cpp structs to indicate whether there's a partocular field set in place or not. With the deprecation of `__isset` field, we want to migrate existing sites of usages to newer API `_ref()` accessors in cpp.

For example, every smiliar to the following:
```
if (s.__isset.x) { ... }
```
should be substituted to
```
if (s.x_ref().has_value()) { ... }
```
if `x` is an unqualified field. When `x` is optional we can drop the memeber call to `has_value()`.

In addition, for writes to `__isset` fields like
```
s.__isset.x = true;
```
we want to migrate it to
```
s.x_ref().ensure(); // if x is unqualified
```
or
```
apache::thrift::ensure_isset_unsafe(s.x_ref()); // if x is optional
```

drop-conflicts

allow-large-files

Reviewed By: snarkmaster

Differential Revision: D22149492

fbshipit-source-id: 3b7f6e7f549e6f0e9c687a22687b7477f7ad107f